refactor: #3 SignUp DTO 수정
